pyproject.toml
[project]
name = "loss-weight-easily"
version = "0.2.0"
description = "🍎 轻松减重助手 - 智能健康管理工具，提供食物查询、体重记录、代谢计算和食谱规划"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "ZDP-Q" }
]
keywords = [
    "food",
    "calories",
    "nutrition",
    "faiss",
    "semantic-search",
    "chinese",
    "multilingual"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database :: Database Engines/Servers",
]

dependencies = [
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "PySide6>=6.6.0",
    "PyYAML>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
loss-weight = "loss_weight.cli:main"
loss-weight-ui = "loss_weight.ui:main"

[project.urls]
Homepage = "https://github.com/ZDP-Q/LossWeightEasily"
Repository = "https://github.com/ZDP-Q/LossWeightEasily"
Issues = "https://github.com/ZDP-Q/LossWeightEasily/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/loss_weight"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    # 忽略 FAISS 的 SwigPy 弃用警告（第三方库问题）
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    # 忽略 pydantic 内部使用的弃用警告
    "ignore:.*pkg_resources.*:DeprecationWarning",
    # 忽略 numpy 相关警告
    "ignore:.*np.float.*:DeprecationWarning",
]
markers = [
    "unit: 单元测试 - 测试单个函数或类",
    "integration: 集成测试 - 测试多个组件协作",
    "e2e: 端到端测试 - 测试完整工作流程",
    "slow: 慢速测试 - 运行时间较长的测试",
    "requires_db: 需要真实数据库 - 需要已初始化的数据库",
    "requires_model: 需要嵌入模型 - 需要下载的模型文件",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["loss_weight"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
